+++ Utilities.h
@@ -8,8 +8,8 @@
 #include <vector>
 #include <string>
 
-int randomChoiceOfN(int n);
-int randomFromXToY(int x, int y);
+long randomChoiceOfN(int n);
+long randomFromXToY(int x, int y);
 float randomFloatZeroToN(float n);
 float randomSampleNormal(float, float);
 float randomSampleNormal(float mean, float dev, float min, float max);
